# Use optimized queue everywhere