* Switch to [Chromium's `SimpleQueue` implementation](https://chromium.googlesource.com/chromium/src/+/0aee4434a4dba42a42abaea9bfbc0cd196a63bc1/third_party/blink/renderer/core/streams/SimpleQueue.js) for (slightly) better performance and `.forEach` support.
* Use `SimpleQueue` for:
  * `ReadableStreamDefaultReader._readRequests`
  * `ReadableStreamBYOBReader._readIntoRequests`
  * `ReadableByteStreamController._pendingPullIntos`
  * `WritableStream._writeRequests`

Benchmark before:
```
> node --prof index.js

web-streams-polyfill testCount(3545) x 742 ops/sec ±2.47% (79 runs sampled) (period: 1.35ms)
web-streams-polyfill testCount(7090) x 362 ops/sec ±1.06% (80 runs sampled) (period: 2.76ms)
web-streams-polyfill testCount(14180) x 167 ops/sec ±1.25% (75 runs sampled) (period: 6.00ms)
web-streams-polyfill testCount(28360) x 73.14 ops/sec ±0.43% (81 runs sampled) (period: 13.67ms)
web-streams-polyfill testCount(56720) x 23.82 ops/sec ±4.17% (57 runs sampled) (period: 41.97ms)
web-streams-polyfill testCount(113440) x 11.08 ops/sec ±0.95% (54 runs sampled) (period: 90.26ms)
```
Benchmark after:
```
> node --prof index.js

web-streams-polyfill testCount(3545) x 972 ops/sec ±1.05% (80 runs sampled) (period: 1.03ms)
web-streams-polyfill testCount(7090) x 460 ops/sec ±1.28% (82 runs sampled) (period: 2.17ms)
web-streams-polyfill testCount(14180) x 209 ops/sec ±1.12% (78 runs sampled) (period: 4.79ms)
web-streams-polyfill testCount(28360) x 87.96 ops/sec ±0.51% (79 runs sampled) (period: 11.37ms)
web-streams-polyfill testCount(56720) x 26.73 ops/sec ±4.14% (63 runs sampled) (period: 37.41ms)
web-streams-polyfill testCount(113440) x 12.54 ops/sec ±0.80% (60 runs sampled) (period: 79.73ms)
```